refactor: 파일 이름 snake-case로 통일
- issue: #1
